+++ CHANGELOG.md
@@ -1,8 +1,16 @@
-# v6.1.1
+# v7.0.1
 
 ## Pass latest model as parameter to Fieldset.onRefreshRequirements
 
 This enables consumers to use this model rather than relying on the two-way bound model in the props, which may not have been updated at the point of the callback triggering.
+
+# v7.0.0
+
+## Fieldsets now remove invalid properties from the model when receiving new fields
+
+BREAKING: If the set of fields is updated, any properties of the model that are not reflected in the new set of fields will be removed. This could be a breaking change if a consumer relies on the fieldset leaving other properties of a model untouched.
+
+Alongside those changes we also improve the checkbox-group implementation to work with numeric enums as well as strings. The requirements service now better identifies checkbox-groups based on their data type, and we have improved the definition list to show the labels for the enums rather than the raw values.
 
 # v6.1.0